refactor: replace double with BigDecimal for monetary values

+++ src/Main.java
@@ -1,17 +1,19 @@
+import java.math.BigDecimal;
+
 public class Main {
     static void main(String[] args) {
         Customer customer = new Customer("Ivan Petrov", "ivan@gmail.com");
 
         Invoice invoice = new Invoice("INV-01", customer);
 
-        invoice.addItem(new InvoiceItem("Website development",1,1200));
-        invoice.addItem(new InvoiceItem("Hosting 1 year", 1, 120));
+        invoice.addItem(new InvoiceItem("Website development",1,BigDecimal.valueOf(1200)));
+        invoice.addItem(new InvoiceItem("Hosting 1 year", 1, BigDecimal.valueOf(120)));
 
         System.out.println("Status: " + invoice.getStatus());
         invoice.send();
         System.out.println("Status: " + invoice.getStatus());
 
-        invoice.pay(500);
+        invoice.pay(BigDecimal.valueOf(500));
         System.out.println("Paid: " + invoice.getPaidAmount());
         System.out.println("Open:" + invoice.getOpenAmount());
         System.out.println("Status:" + invoice.getStatus());
@@ -20,11 +22,13 @@ public class Main {
         invoice.updateOverdueStatus();
         System.out.println("Status after overdue check: " + invoice.getStatus());
 
-        invoice.pay(1000);
+        invoice.pay(BigDecimal.valueOf(1000));
         System.out.println("Status: " + invoice.getStatus());
 
 
         System.out.println("Total invoices: " + invoice.getTotalAmount());
+
+
         //System.out.println(invoice.getStatus());
         //invoice.pay(500);
         //System.out.println(invoice.getStatus());
